requirements.txt
SQLAlchemy~=2.0.35
fastapi~=0.115.0
uvicorn~=0.31.0
pydantic~=2.9.2
pytest~=8.3.3
pytest-asyncio~=0.24.0
python-dotenv~=1.0.1
httpx
